+++ NEWS.md
@@ -1,5 +1,7 @@
 # CHANGES IN xfun VERSION 0.44
 
+- Exported the internal function `md5()` to calculate the MD5 checksums of R objects. The function is essentially a workaround for `tools::md5sum()` (see HenrikBengtsson/Wishlist-for-R#21).
+
 - `write_utf8()` returns the `con` argument (typically a file path) now. Previously, it returns `NULL`.
 
 - Fixed an edge case in `prose_index()`, in which inline code was incorrectly recognized as a code block fence.